#2033 Split access_info test into two parts, according to the files splitting done.
